+++ packages/gamut/src/Button/ButtonInner.tsx
@@ -1,4 +1,4 @@
-import { fontSize, timing } from '@codecademy/gamut-styles';
+import { fontSize, fontSmoothing, timing } from '@codecademy/gamut-styles';
 import isPropValid from '@emotion/is-prop-valid';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
@@ -12,5 +12,7 @@ export const ButtonInner = styled('span', {
     font-size: ${fontSize[16]};
     transition: ${timing.fast} background-color, ${timing.fast} box-shadow,
       ${timing.fast} color;
+
+    ${fontSmoothing()}
   `;
 });